refactor(tui): address low-priority code quality items from TODO.md
Code Style improvements:
- Extract duplicated `buffer_to_string` test helper to shared `test_utils` module
- Extract duplicated layout constants (HEADER_HEIGHT, TASK_CARD_HEIGHT) to `layout` module
- Document `Lane::new` const behavior (Rust 2024 const Vec::new())

Safety/Robustness improvements:
- Document `install_panic_hook` behavior (hook chaining, order of operations)
- Add safe `selected_lane_ref()` helper with explicit panic message for invariant
- Use `.get()` guard in click handler for external input validation

JJ-Change-Id: pqolmu
