+++ crates/tui/src/widgets/board.rs
@@ -107,6 +107,7 @@ pub fn render_board(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::test_utils::buffer_to_string;
     use whip_protocol::Task;
 
     #[test]
@@ -148,19 +149,5 @@ mod tests {
 
         // Should not panic with narrow area
         render_board(&board, 0, None, area, &mut buf);
-    }
-
-    /// Helper to convert buffer to string for testing.
-    fn buffer_to_string(buf: &Buffer) -> String {
-        let mut result = String::new();
-        for y in 0..buf.area.height {
-            for x in 0..buf.area.width {
-                if let Some(cell) = buf.cell((x, y)) {
-                    result.push_str(cell.symbol());
-                }
-            }
-            result.push('\n');
-        }
-        result
     }
 }
